fix: rename email.py to email_utils.py to avoid shadowing stdlib
email.py shadows Python's built-in `email` package, causing
`ModuleNotFoundError: No module named 'email.message'` when aiohttp
(or any other library) tries to import from the stdlib email package.
This breaks standalone execution outside the HA structure.
